Update tests for the GamePageController -> PuzzlesPageController merge
Deleted GamePageControllerTest.scala and GamePageRoutesTest.scala. Merged
the still-relevant iframe-slug test cases into PuzzlesPageControllerTest
(renderGame/renderGameJson: happy path via a mocked DotcomRenderingService,
and not-found for an unrecognised slug) and PuzzlesRoutesTest (route
reversal for the Game Page HTML/JSON endpoints, plus a check that they
don't clash with the existing crossword routes). All crossword-specific
test cases (real CAPI crossword fetch, crossword JSON field assertions,
the display-date formatting test) are gone entirely - nothing
crossword-related survives in these test names/content, matching the
controller changes.

Removed the now-deleted GamePageControllerTest/GamePageRoutesTest entries
from applications/test/package.scala's ApplicationsTestSuite registration.

Full applications/testOnly test.ApplicationsTestSuite: 141/141 passing
(down from 147, as expected - the removed crossword-specific test cases
account for the difference).
