+++ applications/test/package.scala
@@ -40,8 +40,6 @@ class ApplicationsTestSuite
       new PuzzlesPageControllerTest,
       new PuzzlesLayoutProviderTest,
       new PuzzlesRoutesTest,
-      new GamePageControllerTest,
-      new GamePageRoutesTest,
       new IndexPageTest,
       new InteractivePickerTest,
     )
